Pulled out the getting of the set of questions into its own function to allow it to be called multiple times. Function is called getQuestions().
The retakeQuiz button wasn't working after taking a quiz for the second time due to multiple  fires of the eventListener. Spent some time researching a way to properly handle adding and remove eventListeners and found that calling add and removeEventListener works when the events match exactly. This doesn't work when the event has an anonymous function since no two anonymous functions are the same so I took the anonymous functions and put them into their own function to reference multiple times. This was only done for the close and retake button and are added and removed depending on if they are displayed on the page, using the displayQuizResultsHTML and displayQuizHTML. Might be something to work on for the other buttons.